Restructure files to prepare for new list types
