+++ .changeset/thin-cycles-fix.md
@@ -1,0 +1,5 @@
+---
+'@keystonejs/keystone': patch
+---
+
+Restructured internal files.
