+++ packages/keystone/lib/ListTypes/index.js
@@ -1,0 +1,3 @@
+const List = require('./list');
+
+module.exports = { List };
